Move ConnectionHandler & RequestHandler into their own modules
